Add 'purge_after_run' setting for acceptance
This new setting allows us to flag whether or
not to remove packages and do other exhaustive
cleanup after a run.  The previous behavior
was to always do this, which was useful when
doing repeated runs of the acceptance tests on
a VM without restoring a snapshot of the VM
between runs.  This commit changes the default
behavior so that the packages are *not* removed
after a test run, which is more useful for
debugging the runs on CI; the original behavior
can still be achieved by defining the setting
via an environment variable, etc.

+++ acceptance/setup/post_suite/99_teardown.rb
@@ -19,13 +19,14 @@ step "Stop puppetdb" do
   stop_puppetdb(database)
 end
 
-
-if (test_config[:install_type] == :package)
-  step "Uninstall packages" do
-    uninstall_package(database, test_config[:os_families], "puppetdb")
-    uninstall_package(master, test_config[:os_families], "puppetdb-terminus")
-    hosts.each do |host|
-      uninstall_package(host, test_config[:os_families], "puppet")
+if (test_config[:purge_after_run])
+  if (test_config[:install_type] == :package)
+    step "Uninstall packages" do
+      uninstall_package(database, test_config[:os_families], "puppetdb")
+      uninstall_package(master, test_config[:os_families], "puppetdb-terminus")
+      hosts.each do |host|
+        uninstall_package(host, test_config[:os_families], "puppet")
+      end
     end
   end
 end
